requirements.txt
functions-framework==3.0.0
requests==2.25.1
flask==2.0.2
python-dotenv==0.19.2
